refactor: ギア情報サマリを GearInfoSummary 共通コンポーネントに抽出
カード展開時の 5 ブロック（name / category+type / quantity+weight / priority+price / season）を
再利用可能な GearInfoSummary コンポーネントに切り出し。
テーブル列順の情報表示ロジックと PRIORITY_STYLE を一箇所に集約し、
将来の他所（詳細モーダル、共有カード等）からも使い回せるように。
